wrong
something went wrong

+++ Src/globalDebug.c
@@ -3,5 +3,5 @@
 
 #ifdef DEBUG_ROTARY
 volatile GPIO_PinState debugRotaryChanel_1 = GPIO_PIN_RESET, debugRotaryChanel_2 = GPIO_PIN_RESET;
-volatile int32_t debugRotaryPosition = 0;
+volatile int16_t debugRotaryPosition = 0;
 #endif
